+++ src/index.ts
@@ -1,12 +1,11 @@
 #!/usr/bin/env node
 
 import ICommand from './Command/ICommand';
-import * as parameters from '../config/parameters';
 import { applet } from './Applet/appletCommand';
 import { login } from './Auth/loginCommand';
 import { organization } from './Organization/organizationCommand';
 import { timing } from './Timing/timingCommand';
-import { processCommand } from './Command/commandProcessor';
+import { processCommand, API_URL_OPTION } from './Command/commandProcessor';
 
 const index: ICommand = {
 	name: 'sos',
@@ -22,7 +21,7 @@ const index: ICommand = {
 			description: `(default) Command name`,
 		},
 		{ name: 'help', alias: 'h', type: Boolean, description: 'Display this usage guide.' },
-		{ name: 'api-url', alias: 'u', type: String, defaultValue: parameters.apiUrl, description: 'API URL to be used for REST requests' },
+		API_URL_OPTION,
 	],
 	commands: [
 		applet,