+++ config/parameters.js
@@ -2,7 +2,7 @@
 const path = require('path');
 const packageConfig = require('../package.json');
 const environment = process.env.NODE_ENV || 'dev';
-const rootPath = __dirname + path.normalize('/..');
+const rootPath = path.normalize(__dirname + '/..');
 const testsPath = rootPath + '/tests';
 const distPath = rootPath + '/dist';