+++ CHANGELOG.md
@@ -8,7 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 ### Added
 - Login account using username/email and password to access other REST resources `sos login`
-- Organization listing of currently logged account `sos organization list`
+- Organization listing of currently logged account `sos organization list` & `sos organization get`
+- Timing listing of specific device and organization `sos timing list`
+
+### Fixed
+- New UI for `--help` guide
 
 ## [0.1.0] - 2019-08-02
 ### Added