+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.2.0] - 2019-09-05
+## [Unreleased]
 ### Added
 - Login account using username/email and password to access other REST resources `sos login`
 - Organization listing of currently logged account `sos organization list` & `sos organization get`
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - New UI for `--help` guide
 - `--api-url` will change the base url for REST API
 - .env file is looked for in default location first
+- Publishing public npm registry
 
 ## [0.1.0] - 2019-08-02
 ### Added